tests/python/requirements.txt
# Core testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-timeout>=2.2.0

# Async support
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0

# Mocking
pytest-mock>=3.12.0
respx>=0.20.0  # For mocking HTTPX

# Type checking (runtime)
typing-extensions>=4.8.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Video processing (for integration tests)
Pillow>=10.0.0  # For creating test images

# Development
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Linting
mypy>=1.7.0    # Type checking

# Documentation
pytest-html>=4.1.0  # HTML reports
